requirements.txt
Django==5.2.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.8
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-decouple==3.8
python-dateutil==2.9.0.post0
unidecode==1.3.8
requests==2.32.3
django-cors-headers==4.3.1